Modernize helpers actions
