+++ src/redux/helpers/types.ts
@@ -1,0 +1,8 @@
+export enum Providers {
+  'https://accounts.google.com' = 'https://accounts.google.com',
+  'google.com' = 'google.com',
+  'https://www.facebook.com' = 'https://www.facebook.com',
+  'facebook.com' = 'facebook.com',
+  'https://twitter.com' = 'https://twitter.com',
+  'twitter.com' = 'twitter.com',
+}
